distributing methods by class, adding statistics, adding file parsing

+++ src/main/java/org/example/Type.java
@@ -2,6 +2,6 @@ package org.example;
 
 public enum Type {
     STRING,
-    INTEGER,
+    LONG,
     DOUBLE
 }
